+++ src/components/gherkin/attachment/base64Decode.ts
@@ -1,0 +1,4 @@
+export function base64Decode(body: string) {
+  // TODO handle unicode
+  return atob(body)
+}