+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Support for externalised attachments ([#353](https://github.com/cucumber/react-components/pull/353))
 
 ## [22.1.0] - 2024-03-15
 ### Added